Integrated virtual thread support
