+++ antublue-test-engine.properties
@@ -1,3 +1,5 @@
+antublue.test.engine.thread.type=virtual
+#antublue.test.engine.thread.type=platform
 #antublue.test.engine.thread.count=
 #antublue.test.engine.test.class.shuffle=true
 #antublue.test.engine.test.class.include.regex=
